% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch.download.daymet.r
\name{batch.download.daymet}
\alias{batch.download.daymet}
\title{Function to batch download single location DAYMET data}
\usage{
batch.download.daymet(file_location, start_yr = 1980,
  end_yr = as.numeric(format(Sys.time(), "\%Y")) - 1, internal = FALSE)
}
\arguments{
\item{file_location}{: file with several site locations and coordinates
in a format site, latitude, longitude}

\item{start_yr}{: start of the range of years over which to download data}

\item{end_yr}{: end of the range of years over which to download data}

\item{internal}{: TRUE or FALSE, load data into workspace or save to disc}
}
\description{
This function downloads DAYMET data for several single pixel
location.
}
\examples{
batch.download.daymet("mylocation.csv",start_yr=1980,end_yr=2000,internal=TRUE)
}
\keyword{DAYMET,}
\keyword{climate}
\keyword{data}

