% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daymet_grid_agg.r
\name{daymet_grid_agg}
\alias{daymet_grid_agg}
\title{Aggregate daily Daymet data}
\usage{
daymet_grid_agg(file = NULL, int = NULL, FUN = NULL, internal = FALSE,
  output_directory = NULL)
}
\arguments{
\item{file}{character The name of the file to be processed. 
Use daily Daymet data.}

\item{int}{Interval to aggregate by. Options are "monthly", 
"seasonal" or "annual".}

\item{FUN}{Function to be used to aggregate data. Genertic R 
functions can be used. "mean" and "sum" are suggested. na.rm 
= TRUE by default.}

\item{internal}{logical If FALSE, write the output to a tif 
file using the Daymet file format protocol.}

\item{output_directory}{vector (optional) A path to a 
directory where output files should be written. Used only if 
internal = FALSE.}
}
\description{
Aggregates daily Daymet data by time to create 
convenient datasets for data exploration or modelling.
}
\keyword{climate}
\keyword{data}
\keyword{daymet,}
