% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daymet_subset.r
\name{daymet_subset}
\alias{daymet_subset}
\title{Returns an offset dataset with data running from offset DOY in year - 1 to
offset DOY in the current year. Two years of data (730 data layers) are
required for this function to work.}
\usage{
daymet_subset(data, offset = 264)
}
\arguments{
\item{data}{rasterStack or rasterBrick of 730 layers (2 consecutive years)}

\item{offset}{offset of the time series in DOY (default = 264, sept 21)}
}
\description{
The output serves as input for further data processing and
modelling efforts.
}
\examples{

\dontrun{
my_subset = daymet_subset(mystack, offset = 264)
}
}
\keyword{daily}
\keyword{mean}
\keyword{modelling,}
\keyword{temperature}
