% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daymet_tile_tmean.r
\name{daymet_tile_tmean}
\alias{daymet_tile_tmean}
\title{Averages tmax and tmin 'Daymet' tiles into
a single mean daily temperature (tmean) tile
for easy post processing and modelling.}
\usage{
daymet_tile_tmean(path = ".", tile = NULL, year = NULL,
  internal = FALSE)
}
\arguments{
\item{path}{full path location of the daymet tiles}

\item{tile}{which tile to process}

\item{year}{which year to process}

\item{internal}{TRUE / FALSE (if FALSE, write the output to file)
using the Daymet file format protocol.}
}
\description{
Outputs a geotiff using the same naming scheme
as the original files. Optionally a raster object
is returned to the current workspace.
}
\examples{

\dontrun{
# This code calculates the mean temperature
# for all daymet tiles in a user provided
# directory. In this example we first
# download tile 11935 for tmin and tmax

# download a tile
download_daymet_tiles(tiles = 11935,
                      start = 1980,
                      end = 1980,
                      param = c("tmin","tmax"),
                      path = "path_with_daymet_tiles")

# calculate the mean temperature and export
# the result to the R workspace (internal = TRUE)
# If internal = FALSE, a file tmean_11935_1980.tif
# is written into the source path (path_with_daymet_tiles)
tmean = daymet_tile_tmean(path = "path_with_daymet_tiles",
                          tile = 11935,
                          year = 1980,
                          internal = TRUE)
}
}
\keyword{daily}
\keyword{mean}
\keyword{modelling,}
\keyword{temperature}
