% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daymetr.r
\docType{package}
\name{daymetr}
\alias{daymetr}
\alias{daymetr-package}
\title{Interface to the 'Daymet' Web Services}
\description{
Programmatic interface to the 'Daymet' web services (\url{https://daymet.ornl.gov}). Allows for easy downloads of Daymet climate data directly to your R workspace or your computer. Routines for both single pixel data downloads and gridded (netCDF) data are provided.
}
\section{daymetr functions}{

\tabular{llllll}{
\code{\link{download_daymet}}\cr
\code{\link{download_daymet_batch}}\cr
\code{\link{download_daymet_tiles}}\cr
\code{\link{download_daymet_ncss}}\cr
\code{\link{daymet_grid_tmean}}\cr
\code{\link{daymet_grid_offset}}\cr
\code{\link{daymet_grid_agg}}\cr
\code{\link{calc_nd}}
}
}

\author{
Koen Hufkens
}
\keyword{package}
