% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.daymet.r
\name{download.daymet}
\alias{download.daymet}
\title{Function to download single location DAYMET data}
\usage{
download.daymet(site = "Daymet", lat = 36.0133, lon = -84.2625,
  start_yr = 1980, end_yr = as.numeric(format(Sys.time(), "\%Y")) - 1,
  internal = FALSE, quiet = FALSE)
}
\arguments{
\item{site}{: the site name.}

\item{lat}{: latitude (decimal degrees)}

\item{lon}{: longitude (decimal degrees)}

\item{start_yr}{: start of the range of years over which to download data}

\item{end_yr}{: end of the range of years over which to download data}

\item{internal}{: TRUE or FALSE, load data into workspace or save to disc}

\item{quiet:}{TRUE or FALSE, to provide verbose output}
}
\description{
This function downloads DAYMET data for a single
location.
}
\examples{

# NOT RUN
# download.daymet("testsite",
#                 lat=36.0133,
#                 lon=-84.2625,
#                 start_yr=1980)
}
\keyword{DAYMET,}
\keyword{climate}
\keyword{data}

