% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.daymet.tiles.r
\name{download.daymet.tiles}
\alias{download.daymet.tiles}
\title{Function to batch download gridded DAYMET data}
\usage{
download.daymet.tiles(lat1 = 35.6737, lon1 = -86.3968, lat2 = NA,
  lon2 = NA, start_yr = 1980, end_yr = 1980, param = "ALL")
}
\arguments{
\item{lat1}{: top left latitude (decimal degrees)}

\item{lon1}{: top left longitude (decimal degrees)}

\item{lat2}{: bottom right latitude (decimal degrees)}

\item{lon2}{: bottom right longitude(decimal degrees)}

\item{start_yr}{: start of the range of years over which to download data}

\item{end_yr}{: end of the range of years over which to download data}

\item{param}{: climate variable you want to download vapour pressure (vp), 
minimum and maximum temperature (tmin,tmax), snow water equivalent (swe), 
solar radiation (srad), precipitation (prcp) , day length (dayl).
The default setting is ALL, this will download all the previously mentioned
climate variables.}
}
\description{
This function downloads DAYMET data
}
\examples{

# NOT RUN
# download.daymet.tiles(lat1=35.6737,
#                       lon1=-86.3968,
#                       start_yr=1980,
#                       end_yr=1980,
#                       param="ALL")
}
\keyword{climate}
\keyword{data}
\keyword{daymet,}

