% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_daymet.r
\name{download_daymet}
\alias{download_daymet}
\title{Function to download single location DAYMET data}
\usage{
download_daymet(site = "Daymet", lat = 36.0133, lon = -84.2625,
  start = 2000, end = as.numeric(format(Sys.time(), "\%Y")) - 1,
  internal = FALSE, quiet = FALSE)
}
\arguments{
\item{site}{: the site name.}

\item{lat}{: latitude (decimal degrees)}

\item{lon}{: longitude (decimal degrees)}

\item{start}{: start of the range of years over which to download data}

\item{end}{: end of the range of years over which to download data}

\item{internal}{: takes FALSE, "assign" or "data.frame",
Download to file (FALSE) or "assign" a variable dynamically or return a
data frame to the command line}

\item{quiet:}{TRUE or FALSE, to provide verbose output}
}
\description{
This function downloads DAYMET data for a single
location.
}
\examples{

\dontrun{
download_daymet("testsite_name",
                lat=36.0133,
                lon=-84.2625,
                start=2000)
}
}
\keyword{DAYMET,}
\keyword{climate}
\keyword{data}
