% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_daymet_batch.r
\name{download_daymet_batch}
\alias{download_daymet_batch}
\title{Function to batch download single location DAYMET data}
\usage{
download_daymet_batch(file_location, start_yr = 1980,
  end_yr = as.numeric(format(Sys.time(), "\%Y")) - 1, internal = "assign")
}
\arguments{
\item{file_location}{: file with several site locations and coordinates
in a format site, latitude, longitude}

\item{start_yr}{: start of the range of years over which to download data}

\item{end_yr}{: end of the range of years over which to download data}

\item{internal}{: assign or FALSE, load data into workspace or save to disc}
}
\description{
This function downloads DAYMET data for several single pixel
location.
}
\examples{

# NOT RUN
# download_daymet_batch("yourlocations.csv")
}
\keyword{DAYMET,}
\keyword{climate}
\keyword{data}
