% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_daymet_ncss.r
\name{download_daymet_ncss}
\alias{download_daymet_ncss}
\title{Function to geographically subset regions exceeding tile limits}
\usage{
download_daymet_ncss(location = c(36.61, -85.37, -81.29, 33.57),
  start = 1988, end = 1988, param = "tmin", path = ".")
}
\arguments{
\item{location}{location of a bounding box c(lat, lon, lat, lon) defined
by a top left and bottom-right coordinates}

\item{start}{start of the range of years over which to download data}

\item{end}{end of the range of years over which to download data}

\item{param}{climate variable you want to download vapour pressure (vp), 
minimum and maximum temperature (tmin,tmax), snow water equivalent (swe), 
solar radiation (srad), precipitation (prcp) , day length (dayl).
The default setting is ALL, this will download all the previously mentioned
climate variables.}

\item{path}{directory where to store the downloaded data}
}
\value{
netCDF data file of an area circumscribed by the location bounding
box
}
\description{
This function downloads DAYMET data
}
\examples{

\dontrun{
download_daymet_ncss(location = c(36.61,-85.37,-81.29,33.57),
                      start = 1980,
                      end = 1980,
                      param = "tmin")
}
}
\keyword{climate}
\keyword{data}
\keyword{daymet,}
