% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get.Daymet}
\alias{get.Daymet}
\title{Function to download single location DAYMET data}
\usage{
get.Daymet(site = "Daymet", lat = 36.0133, lon = -84.2625,
  start_yr = 1980, end_yr = as.numeric(format(Sys.time(), "\%Y")) - 2,
  internal = TRUE)
}
\arguments{
\item{site}{: the site name.}

\item{lat}{: latitude (decimal degrees)}

\item{lon}{: longitude (decimal degrees)}

\item{start_yr}{: start of the range of years over which to download data}

\item{end_yr}{: end of the range of years over which to download data}

\item{internal}{: TRUE or FALSE, load data into workspace or save to disc}
}
\description{
This function downloads DAYMET data for a single
location.
}
\examples{
get.Daymet("testsite",lat=36.0133,lon=-84.2625,start_yr=1980,end_yr=2000)
}
\keyword{DAYMET,}
\keyword{climate}
\keyword{data}

