% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc2tif.r
\name{nc2tif}
\alias{nc2tif}
\title{Converts all .nc files in a given directory to .tif
and preserves the naming conventions of the files that
are converted.}
\usage{
nc2tif(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{a character string showing the path to the 
directory containing Daymet .nc files}

\item{overwrite}{a logical controlling whether all 
files will be written, or whether files will not be 
written in the event that there is already a .tif of 
that file.}
}
\description{
Conversion to .tif may simplify workflows if the data
that has been downloaded is to be handled in other
software (e.g. QGIS).
}
