% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tile_outlines}
\alias{tile_outlines}
\title{tile_outlines}
\format{
SpatialPolygonDataFrame
\describe{
  \item{TileID}{tile ID number}
  \item{XMin}{minimum longitude}
  \item{XMax}{maximum longitude}
  \item{YMin}{minimum latitude}
  \item{YMax}{maximum latitude}
}
}
\usage{
tile_outlines
}
\description{
Large simple feature collection containing the outlines of all the 
Daymet tiles available as well as projection information. This data
was converted from a shapefile as provided on the Daymet main website.
}
\keyword{datasets}
