% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daymet_tile_tmean.r
\name{daymet_tile_tmean}
\alias{daymet_tile_tmean}
\title{Averages tmax and tmin 'Daymet' tiles into
a single mean daily temperature (tmean) tile
for easy post processing and modelling.}
\usage{
daymet_tile_tmean(path = ".", tile = NULL, year = NULL,
  internal = FALSE)
}
\arguments{
\item{path}{full path location of the daymet tiles}

\item{tile}{which tile to process}

\item{year}{which year to process}

\item{internal}{TRUE / FALSE (if FALSE, write the output to file)
using the Daymet file format protocol.}
}
\description{
Outputs a geotiff using the same naming scheme
as the original files. Optionally a raster object
is returned to the current workspace.
}
\examples{

\dontrun{
daymet_tile_tmean(path = ".",
             tile = 11935,
             year = 2015)
}
}
\keyword{daily}
\keyword{mean}
\keyword{modelling,}
\keyword{temperature}
