% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_daymet_batch.r
\name{download_daymet_batch}
\alias{download_daymet_batch}
\title{Function to batch download single location DAYMET data}
\usage{
download_daymet_batch(batch_file, start = 1980,
  end = as.numeric(format(Sys.time(), "\%Y")) - 1, internal = TRUE)
}
\arguments{
\item{batch_file}{file with several site locations and coordinates
in a format site, latitude, longitude}

\item{start}{start of the range of years over which to download data}

\item{end}{end of the range of years over which to download data}

\item{internal}{assign or FALSE, load data into workspace or save to disc}
}
\value{
a nested list or writes data to file
}
\description{
This function downloads DAYMET data for several single pixel
location.
}
\examples{

\dontrun{
download_daymet_batch(file_location = "yourlocations.csv")
}
}
\keyword{DAYMET,}
\keyword{climate}
\keyword{data}
